+++ Business/Models/Movimento.cs
@@ -3,7 +3,7 @@
     internal class Movimento
     {
         // Construtor
-        public Movimento(DateTime data, string descricao, int valor, string marcacao, int clienteId)
+        public Movimento(DateTime data, string descricao, decimal valor, string marcacao, int clienteId)
         {
             this.Data = data;
             this.Descricao = descricao;
@@ -16,16 +16,12 @@
         public int Id { get; set; }
         
 
-        public DateTime Data { get; set; }
-
+        public DateTime Data { get; set; } 
         public string? Descricao { get; set; }
-        public int Valor { get; set; }
+        public decimal Valor { get; set; }
         public char Tipo { get; set; }
-
         public string? Marcacao { get; set; }
-
         public int ClienteId { get; set; }
-
         public Cliente Cliente { get; set; } = null!;
 
     }
